+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 1.3-SNAPSHOT
 
+- Implementing bulk deserialization of arrays
 - Adding an API call to get a message on a topic at a specific index
 
 1.2